test: split real-LS-data Diagram test into capture(file)+render(snapshot)
Rework the real-rpc-data test to the agreed shape:
 - capture (nightly, real LS): Diagram.flowModel.capture.test.ts fetches
   getFlowModel and writes/drift-checks fixtures/realdata/flowModel.json —
   the rpc-contract assertion, L3-tier, off the PR path.
 - render (fast PR, no LS): Diagram.flowModel.render.test.tsx loads that
   committed real flow model, renders the Diagram, and snapshots it (+ a
   semantic assert). jsdom, no distro.

Replaces the earlier combined live fetch-and-render test. Fast config runs
the render/snapshot and excludes *.capture.test.ts; capture runs via
jest.realdata.config.js nightly.
